# opaque draw fastpath, simd mask magnifyBy2, float issue fix for 360deg rotation