--- a/src/pixie/internal.nim
+++ b/src/pixie/internal.nim
@@ -96,6 +96,31 @@ proc toPremultipliedAlpha*(data: var seq[ColorRGBA | ColorRGBX]) {.raises: [].} 
       c.b = ((c.b.uint32 * c.a.uint32) div 255).uint8
       data[j] = c
 
+proc isOpaque*(data: var seq[ColorRGBX], start, len: int): bool =
+  result = true
+
+  var i: int
+  when defined(amd64) and not defined(pixieNoSimd):
+    let
+      vec255 = mm_set1_epi32(cast[int32](uint32.high))
+      colorMask = mm_set1_epi32(cast[int32]([255.uint8, 255, 255, 0]))
+    for _ in 0 ..< len div 16:
+      let
+        values0 = mm_loadu_si128(data[i + 0].addr)
+        values1 = mm_loadu_si128(data[i + 4].addr)
+        values2 = mm_loadu_si128(data[i + 8].addr)
+        values3 = mm_loadu_si128(data[i + 12].addr)
+        values01 = mm_and_si128(values0, values1)
+        values23 = mm_and_si128(values2, values3)
+        values = mm_or_si128(mm_and_si128(values01, values23), colorMask)
+      if mm_movemask_epi8(mm_cmpeq_epi8(values, vec255)) != 0xffff:
+        return false
+      i += 16
+
+  for j in i ..< len:
+    if data[j].a != 255:
+      return false
+
 when defined(amd64) and not defined(pixieNoSimd):
   proc packAlphaValues*(v: M128i): M128i {.inline, raises: [].} =
     ## Shuffle the alpha values for these 4 colors to the first 4 bytes