… tests/fileformats/svg/noto-emoji.png
… tests/fileformats/svg/openmoji.png
… tests/images/rotate360.png